requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-celery==3.2.1
django-debug-toolbar==2.1
django-import-export==1.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
jdcal==1.4.1
kombu==3.0.37
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.1
pytz==2019.3
PyYAML==5.1.2
redis==2.10.6
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
